[FIX] 장바구니에 상품이 없을 때도 주문이 생성되는 오류 수정, 주문 상세 페이지 생성

+++ src/main/java/com/ll/synergarette/boundedContext/order/service/OrderService.java
@@ -1,5 +1,6 @@
 package com.ll.synergarette.boundedContext.order.service;
 
+import com.ll.synergarette.base.rsData.RsData;
 import com.ll.synergarette.boundedContext.cartItem.entity.CartItem;
 import com.ll.synergarette.boundedContext.cartItem.service.CartItemService;
 import com.ll.synergarette.boundedContext.member.entity.Member;
@@ -14,6 +15,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -26,12 +28,20 @@ public class OrderService {
 
 
     @Transactional
-    public Order createFromCart(Member buyer) {
+    public RsData<Order> createFromCart(Member buyer) {
         List<CartItem> cartItemList = cartItemService.getItemsByBuyer(buyer);
+
+        if(cartItemList.isEmpty()){
+            return RsData.of("F-1", "장바구니에 상품이 없습니다.");
+        }
+
 
         List<OrderItem> orderItemList = new ArrayList<>();
 
-        cartItemList.stream().map(CartItem::getGoodsItem).forEach(goods -> orderItemList.add(new OrderItem(goods)));
+        cartItemList
+                .stream()
+                .map(CartItem::getGoodsItem)
+                .forEach(goodsItem -> orderItemList.add(new OrderItem(goodsItem)));
         // 장바구니 상품들을 가져와서 order 아이템 리스트에 담는다.
 
         cartItemList.stream().forEach(cartItem -> cartItemService.deleteCartItem(cartItem));
@@ -40,7 +50,7 @@ public class OrderService {
     }
 
     @Transactional
-    public Order create(Member buyer, List<OrderItem> orderItemList){
+    public RsData<Order> create(Member buyer, List<OrderItem> orderItemList){
         Order order = Order
                 .builder()
                 .member(buyer)
@@ -52,6 +62,14 @@ public class OrderService {
 
         orderRepository.save(order);
 
-        return order;
+        return RsData.of("S-1", "주문이 성공적으로 생성되었습니다.", order) ;
+    }
+
+    public Optional<Order> findForPrintById(Long id) {
+        return findById(id);
+    }
+
+    public Optional<Order> findById(Long id){
+        return orderRepository.findById(id);
     }
 }
